streamlitdatavisuproject/requirements.txt
streamlit
numpy
pandas
datetime
seaborn
matplotlib.pyplot